I have created a branch called stepOneHatIntro. This branch modifies index.html, main.css, main.js, and README.md
The changes made to index.html was adding a div with a card and another div that is empty to the body. The changes made to main.css was adding .sorting-hat to change the size of the hat image. The changes made to main.js was removing my initial console log, adding an array for a set of objects at the beginning, and adding an event listener to listen for a button click. The changes made to README.md was adding steps in replacement of each dashed point in the instructions section.