I have created a branch called stepTwoRandomHouse. This branch modifies index.html and main.js
The changes made to index.html was formatting the document with prettier code formatter, I also added a div with an id of #placeHere. The changes made to main.js was formatting the file with prettier code formatterr, adding the names of houses to existing students, creating an array of house objects with the idNums and house distinguishing which house is which, adding a name to the input of the name form, adding an onclick to the button on he form so that the future code could run, and finally adding a function that runs a series of functions when the submit button is clicked. Those seriees of functions are displaying current students, having a random house generator, and pushing a newly created student to the current students list.